fixed linking issue after renameing files
